Ai Projject/Ai Projject/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
faiss-cpu==1.7.4
sentence-transformers==3.0.1
langdetect==1.0.9
transformers==4.45.0
torch==2.2.0
numpy==1.26.4
python-dotenv==1.0.0
huggingface-hub>=0.23.2
